refactor CLI
